changing prices from long to decimal for overflow

+++ SimpleHashDotCom/SimpleHashDotCom.Sdk/Models/Nft/Bid.cs
@@ -4,33 +4,33 @@ namespace SimpleHashDotCom.Sdk.Models.Nft;
 
 public class Bid
 {
-	[JsonPropertyName("id")]
-	public string Id { get; set; } = "";
+    [JsonPropertyName("id")]
+    public string Id { get; set; } = "";
 
-	[JsonPropertyName("quantity")]
-	public long? Quantity { get; set; }
+    [JsonPropertyName("quantity")]
+    public long? Quantity { get; set; }
 
-	[JsonPropertyName("price")]
-	public long Price { get; set; }
+    [JsonPropertyName("price")]
+    public decimal Price { get; set; }
 
-	[JsonPropertyName("timestamp")]
-	public DateTime Timestamp { get; set; }
+    [JsonPropertyName("timestamp")]
+    public DateTime Timestamp { get; set; }
 
-	[JsonPropertyName("bidder_address")]
-	public string BidderAddress { get; set; } = "";
+    [JsonPropertyName("bidder_address")]
+    public string BidderAddress { get; set; } = "";
 
-	[JsonPropertyName("marketplace_id")]
-	public string MarketplaceId { get; set; } = "";
+    [JsonPropertyName("marketplace_id")]
+    public string MarketplaceId { get; set; } = "";
 
-	[JsonPropertyName("collection_id")]
-	public string? CollectionId { get; set; }
+    [JsonPropertyName("collection_id")]
+    public string? CollectionId { get; set; }
 
-	[JsonPropertyName("nft_id")]
-	public string NftId { get; set; } = "";
+    [JsonPropertyName("nft_id")]
+    public string NftId { get; set; } = "";
 
-	[JsonPropertyName("payment_token")]
-	public PaymentToken PaymentToken { get; set; } = new();
+    [JsonPropertyName("payment_token")]
+    public PaymentToken PaymentToken { get; set; } = new();
 
-	[JsonPropertyName("expiration_timestamp")]
-	public DateTime? ExpirationTimestamp { get; set; }
+    [JsonPropertyName("expiration_timestamp")]
+    public DateTime? ExpirationTimestamp { get; set; }
 }
